requirements.txt
Pillow ~= 8.3.1
discord.py ~= 1.7.3
python-dotenv ~= 0.18.0
ffprobe-python ~= 1.0.3